requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
celery==5.4.0
redis==5.0.8
django-cors-headers==4.4.0
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
python-dotenv>=1.0
gunicorn>=21.2
requests==2.31.0
Pillow>=10.2
markdown>=3.5
django-celery-results==2.5.1
django-celery-beat==2.6.0
google-api-python-client==2.118.0
pymysql>=1.1.0
